Removed EdonR.hpp Fixed VS.NET 2005 compilation Squelched linker warnings in 2005 and 2008

+++ trunk/tests/ECC_Test/ecc_unit_test.cpp
@@ -838,7 +838,7 @@ void HandshakeTest()
 
 bool got_iv(u64 correct)
 {
-    u32 new_iv_low = correct & AuthenticatedEncryption::IV_MASK;
+    u32 new_iv_low = (u32)(correct & AuthenticatedEncryption::IV_MASK);
 
     static u64 last_iv = 0;
